Use relative imports
This change does not affect to any behaviors of it, but is just stylish
improvements to remove 2to3 dependency.

1. All old-style absolute imports e.g. `from alembic import util` are
   replaced by relative imports e.g. `from . import util`.

2. Use of proxy modules (`alembic.op` and `alembic.context`) are replaced
   by deferred import patterns, but these modules still exist.

3. Imports are grouped by 3 types of their origins (standard libraries,
   third party libraries, and local modules) and blank lines are put between
   each group of imports.  This style is from PEP 8:

   <http://www.python.org/dev/peps/pep-0008/#imports>

For stylish change, I can't do that much, because it was intentionally done
as it was by the author.  If the author will allow me to make it more closer
to PEP 8 style, I will do that as well.