prettier revision identifier (always exatly 12 hex digits long) with better guaranteed entropy (the last 48 bits of a uuid4 are all random). Works everywhere where uuid module exists (2.5 -)
Rationale: with the current code most revision ids start with 1, 2, 3 or 4, and there are far more collisions with 2 initial digits than necessary (cumbersome to check a file by revision on command line and such). So instead of revision ids that are 1-12 digits long with ~45 bits of information, lets just have them 12 digits long and contain full 48 bits of info.