Render autogenerate=True if present
Fixed issue where "autoincrement=True" would not render for a column that
specified it, since as of SQLAlchemy 1.1 this is no longer the default
value for "autoincrement".  Note the behavior only takes effect against the
SQLAlchemy 1.1.0 and higher; for pre-1.1 SQLAlchemy, "autoincrement=True"
does not render as was the case before. Pull request courtesy  Elad Almos.

Change-Id: Ia043c60b4bb6f0520056a08c8d5e2f0c838ddff7
Pull-request: https://bitbucket.org/zzzeek/alembic/pull-requests/80