Update comment documentation, make sure tests always run
Corrected the links and text in the changelog note as well as
ensured new comment-oriented methods and parameters include a
versionadded token.    Added a more specific check so that
any run of the tests will make sure SQLAlchemy issue 4436
is resolved as 1.2.16 resolves it but 1.3.0b1, which currently
comes out for "python setup.py test", does not.

Change-Id: Ibd94bf7940b279e85b0ec1ddceb309df0c18f0b1