Remove formatargspec work from create_method_proxy
Simplified the internal scheme used to generate the ``alembic.op`` namespace
to no longer attempt to generate full method signatures (e.g. rather than
generic ``*args, **kw``) as this was not working in most cases anyway, while
in rare circumstances it would in fact sporadically have access to the real
argument names and then fail when generating the function due to missing
symbols in the argument signature.

Change-Id: I06e4ecf1ddcdb0799856ba7e39b17f28d19cc5d2
Fixes: #548