Ensure default_schema_name set to None in "schema is none" test
A test here is working with the assumption that the default
schema name for the SQLite dialect is None.  This assumption
is likely changing in SQLAlchemy 1.4 due to
Ia5c89eb27cc8dc2c5b8e76d6c07c46290a7901b6, so for this test
case explicitly set it to None.

Change-Id: I8b6427d9f975d489f33f51484e45b0bdc81acf70