render .info in create_table
Fixed regression caused by :ticket:`879` released in 1.7.0 where the
".info" dictionary of ``Table`` would not render in autogenerate create
table statements.  This can be useful for custom create table DDL rendering
schemes so it is restored.

Additionally upon seeing that Rewriter is failing typing that was
just imporved in the previous commit for #1325 /
Ibfb7a57a081818c290cf0964d12a72b85c2c1983, further correct the typing
of the "revision" argument for process_revision_directives which was
still inconsistent.

Change-Id: Ifa4c7bd1b730d51629f42bc159b994f42d157c04
Fixes: #1329