use SQLAlchemy's xdist methods
Fixes to support pytest 8.1 for the test suite.

the use of teardown() was based on pytest's nose
compat, which is removed.  their xdist style tests use the name
"setup_method()" and "teardown_method()" now.

We have SQLAlchemy's pytestplugin in use which uses pytest fixtures
to invoke our own xdist style setUp and tearDown methods, which we
are already using here, so use those for this one test.

Fixes: #1435
Change-Id: I4c49e81fca6bfa957594714009531fe12691ace5