ensure conv is applied to all reflected index and constraint names
The autogenerate process will now apply the :meth:`.Operations.f` modifier
to the names of all constraints and indexes that are reflected from the
target database when generating migrations, which has the effect that these
names will not have any subsequent naming conventions applied to them when
the migration operations proceed.  As reflected objects already include the
exact name that's present in the database, these names should not be
modified.   The fix repairs the issue when using custom naming conventions
which feature the ``%(constraint_name)s`` token would cause names to be
double-processed, leading to errors in migration runs.

Fixes: #264
Change-Id: Ie48f65c41e66f3698cafe558cafa189f524c8cdf