fix(typings): improve typing for server_default
Issue #1669 shows that the typings for server default were inconsistent with sqlaclhemy's definition.

In this PR we:
- fix the definition of `_ServerDefault` to match that of sqlalchemy's `_ServerDefaultArgument`
- use that same definition in all places that use a `server_default` or `existing_server_default`

Note that the last change also changes the default of the argument `existing_server_default` from `False` to `None`. This could be a breaking change, however that argument is currently typed in manyplaces as `Optional[_ServerDefault] = None` (and `_ServerDefault` does not include `bool`).