add test for issue 1743
however this issue looks to be simple sqlite reflection issue
on sqlalchemy side

also fixes the incorrect exclusions code from 20c108720d7127ca.
just skip entire test if SQLite fk reflection is not there.
this exclusion should not have been added originally, it looks
like this was mistaken for a limitation when it really was a bug
in sqlite reflection

Change-Id: I3f51ebaf01f41dba86cea0aa31be28334401887b