force the driver level on the bulk insert PG tests
this changes to use the psycopg2 dialect in all cases
even if the default for postgresql changes.

this is to prepare for SQLAlchemy moving to psycopg as the
default dialect. I had hoped to move these tests to use psycopg
and test for bind casts, however, we are still running tests against
SQLAlchemy 1.4 which does not have the psycopg dialect.   so stick with
psycopg2 for now.

also adds an rmtree for build/ as apparently pip install is re-using
this directory if it's there from a previous build, and the
artfacts left over from the gerrit for "plugins" is interfering with
the stubs generation.   We really should find a way to make the stub
generation / testing more robust as it breaks much too easily and
without an easy way to tell why.

Change-Id: Ibdab43a2cda5ec4f697d35c8a5f090aa56492703